+++ src/layer/tile/TileLayer.Canvas.js
@@ -19,7 +19,7 @@ L.TileLayer.Canvas = L.TileLayer.extend({
 	},
 
 	_redrawTile: function (tile) {
-		this.drawTile(tile, tile._tilePoint, tile._zoom);
+		this.drawTile(tile, tile._tilePoint, this._map._zoom);
 	},
 
 	_createTileProto: function () {
@@ -36,19 +36,18 @@ L.TileLayer.Canvas = L.TileLayer.extend({
 		return tile;
 	},
 
-	_loadTile: function (tile, tilePoint, zoom) {
+	_loadTile: function (tile, tilePoint) {
 		tile._layer = this;
 		tile._tilePoint = tilePoint;
-		tile._zoom = zoom;
-
-		this.drawTile(tile, tilePoint, zoom);
+		
+		this._redrawTile(tile);
 
 		if (!this.options.async) {
 			this.tileDrawn(tile);
 		}
 	},
 
-	drawTile: function (tile, tilePoint, zoom) {
+	drawTile: function (tile, tilePoint) {
 		// override with rendering code
 	},